+++ packages/openapi_base/lib/src/http_headers.dart
@@ -1,5 +1,6 @@
 class OpenApiHttpHeaders {
   static const contentType = 'content-type';
+  static const userAgent = 'user-agent';
 }
 
 class OpenApiHttpStatus {
@@ -9,6 +10,7 @@ class OpenApiHttpStatus {
   static const movedPermanently = 301;
   static const found = 302;
 
+  static const badRequest = 400;
   static const unauthorized = 401;
   static const notFound = 404;
   static const conflict = 409;