+++ packages/openapi_base/lib/src/openapi_base.dart
@@ -1,6 +1,8 @@
 import 'dart:typed_data';
 
+import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
+import 'package:openapi_base/openapi_base.dart';
 import 'package:openapi_base/src/openapi_client_base.dart';
 import 'package:openapi_base/src/openapi_content_type.dart';
 import 'package:uri/uri.dart';
@@ -120,7 +122,7 @@ abstract class OpenApiServerRouterBase {
       List<String>? value,
       required T Function(List<String> value) decode}) {
     if (value == null || value.isEmpty) {
-      throw StateError('Parameter "$name" is required.');
+      throw MissingParameterException(name);
     }
     return decode(value);
   }